Reorganized the location of unit tests to make more sense.
